requirements.txt
alabaster==0.7.7
amqp==1.4.9
anyjson==0.3.3
Babel==2.2.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
BeautifulSoup==3.2.1
billiard==3.3.0.23
celery==3.1.18
certifi==2016.2.28
cffi==1.5.2
click==6.3
contextlib2==0.5.1
coverage==4.0.3
cryptography==1.2.3
cssselect==0.9.1
cssutils==0.9.10
Django==1.6.11
django-bitfield==1.7.1
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.2
django-jsonfield==0.9.13
django-paging==0.2.5
django-picklefield==0.3.2
django-recaptcha==1.0.5
django-rest-swagger==0.3.5
django-social-auth==0.7.28
django-sudo==1.2.1
django-templatetag-sugar==1.0
djangorestframework==2.3.14
docutils==0.12
drfdocs==0.0.9
email-reply-parser==0.2.0
enum34==1.1.2
exam==0.10.5
flake8==2.5.1
hiredis==0.1.6
httplib2==0.9.2
idna==2.0
ipaddr==2.1.11
ipaddress==1.0.16
Jinja2==2.8
kafka==1.0.2
kombu==3.0.30
livereload==2.4.1
lxml==3.5.0
Markdown==2.6.6
MarkupSafe==0.23
mccabe==0.3.1
mkdocs==0.15.3
mkdocs-bootstrap==0.1.1
mkdocs-bootswatch==0.4.0
mock==1.0.1
module==0.2.1
msgpack-python==0.4.7
ndg-httpsclient==0.4.0
oauth2==1.9.0.post1
pep8==1.7.0
petname==1.7
progressbar==2.3
psycopg2==2.6.1
py==1.4.31
pyasn1==0.1.9
pycparser==2.14
pyflakes==1.0.0
Pygments==2.1.3
pyOpenSSL==0.15.1
pytest==2.6.4
pytest-django==2.9.1
python-dateutil==2.5.0
python-memcached==1.57
python-openid==2.2.5
python-pcre==0.7
pytz==2016.1
PyYAML==3.11
raven==5.11.1
rb==1.4
redis==2.10.5
requests==2.9.1
sentry==8.3.0.dev0
setproctitle==1.1.9
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
South==1.0.1
Sphinx==1.3.6
sphinx-rtd-theme==0.1.9
sqlparse==0.1.19
statsd==3.1
tarantool==0.5.4
tornado==4.3
toronado==0.0.7
ua-parser==0.6.1
urllib3==1.14
uWSGI==2.0.12
wheel==0.24.0
xpinyin==0.5.4
